Added aardvark test to demo splitOn usage
Added flag for dontExecuteSql so I can see the db scripts to run without them being executed.  I should consider removing all of this sql setup DDL script code as doing it via sql scripts to be executed via managemant studio is going to simplify things.

Added sql scripts for demo.  Note aardvark script is currently a copy of the dodo script but with the extra stored proc creation so it can be used as a template for the aardvark tests